Updated time-date format
